+++ src/components/common/ConfirmationModal.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import {
+  Box,
   Dialog,
   DialogActions,
   DialogContent,
@@ -12,6 +13,7 @@ export interface ModalProps {
   open: boolean
   title: string
   message: string
+  children?: React.ReactNode
   handleConfirm: () => void
   handleCancel: () => void
 }
@@ -20,6 +22,7 @@ const ConfirmationModal: React.FC<ModalProps> = ({
   open,
   title,
   message,
+  children,
   handleConfirm,
   handleCancel
 }) => {
@@ -35,6 +38,7 @@ const ConfirmationModal: React.FC<ModalProps> = ({
         <DialogContentText id="alert-dialog-description">
           {message}
         </DialogContentText>
+        {children && <Box sx={{ mt: 2 }}>{children}</Box>}
       </DialogContent>
       <DialogActions>
         <Button variant="contained" onClick={handleCancel} color="primary">