+++ src/pages/Mail/Mail-styles.ts
@@ -30,7 +30,8 @@ export const MailBody = styled(Box)(({ theme }) => ({
   display: "flex",
   flexDirection: "row",
   width: "100%",
-  height: "100%",
+  flex: 1,
+  minHeight: 0,
   position: 'relative'
   // overflow: 'auto !important'
 }));